Agentic-ICU-Decision-Support-A-Multi-Agent-LLM-Framework-for-Real-Time-Critical-Care/requirements.txt
# Essential packages for ICU simulation
pandas==2.0.3
numpy==1.24.3
fastapi==0.103.0
uvicorn==0.23.0
pydantic==2.3.0
python-dotenv==1.0.0
faker==19.6.0
